Added remains of the dissected HiGlassComponentTests
