+++ test/utils/empty-body.js
@@ -1,0 +1,6 @@
+export default function emptyBody() {
+  const myNode = global.document.body;
+  while (myNode.lastElementChild) {
+    myNode.removeChild(myNode.lastElementChild);
+  }
+}
